feat:5.16. Модуляция
